requirements.txt
Flask==2.0.1
python-socketio==5.4.0
flask-socketio==5.1.1
psycopg2-binary==2.9.3
python-dotenv==0.19.1
Flask-Cors==3.0.10
Werkzeug==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
gunicorn==20.1.0
gevent==21.12.0
gevent-websocket==0.10.1
